Move and rename all utils modules to ensure consistent names
