+++ __tests__/spec/install.js
@@ -3,7 +3,7 @@
 const fs = require('fs')
 const path = require('path')
 
-const utils = require('../util')
+const utils = require('../utils')
 const { exec } = require('../../lib/exec')
 
 describe('npm install', () => {
