+++ lib/config.js
@@ -6,7 +6,7 @@
 const fse = require('fs-extra')
 const path = require('path')
 
-const { appDir } = require('./path-utils')
+const { appDir } = require('./utils/paths')
 const { isString } = require('lodash')
 const appConfigPath = path.join(appDir, 'config.json')
