+++ lib/config.test.js
@@ -3,7 +3,7 @@
 const fse = require('fs-extra')
 const path = require('path')
 
-const { appDir } = require('./path-utils')
+const { appDir } = require('./utils/paths')
 const appConfig = path.join(appDir, 'config.json')
 require('./config')
 const config = require('./config')
