+++ lib/migrator/file-helpers.spec.js
@@ -5,9 +5,8 @@ const path = require('path')
 
 jest.mock('./logger')
 
-const { mockFileSystem } = require('../../__tests__/util/mock-file-system')
-const { projectDir } = require('../path-utils')
-
+const { mockFileSystem } = require('../../__tests__/utils/mock-file-system')
+const { projectDir } = require('../utils/paths')
 const { getFileAsLines, removeLineFromFile, replaceStartOfFile, deleteFile, deleteDirectory } = require('./file-helpers')
 
 describe('file helpers', () => {
