+++ lib/migrator/index.js
@@ -18,7 +18,7 @@ const {
 const fs = require('fs-extra')
 const path = require('path')
 const { npmInstall, packageJsonFormat } = require('../../bin/utils')
-const { projectDir } = require('../path-utils')
+const { projectDir } = require('../utils/paths')
 
 const minimumPrototypeVersionForMigration = 8
