+++ lib/migrator/logger.js
@@ -1,7 +1,7 @@
 const os = require('os')
 const path = require('path')
 
-const { projectDir } = require('../path-utils')
+const { projectDir } = require('../utils/paths')
 const fs = require('fs').promises
 
 const packageVersion = require('../../package.json').version
