+++ lib/migrator/migration-steps.js
@@ -1,7 +1,7 @@
 const path = require('path')
 const lodash = require('lodash')
 const fse = require('fs-extra')
-const { appDir, projectDir, packageDir } = require('../path-utils')
+const { appDir, projectDir, packageDir } = require('../utils/paths')
 const config = require('../config')
 const logger = require('./logger')
 const { addReporter, reportSuccess, reportFailure } = require('./reporter')
