+++ lib/migrator/migration-steps.spec.js
@@ -39,7 +39,7 @@ const fse = require('fs-extra')
 const reporter = require('./reporter')
 const fileHelpers = require('./file-helpers')
 const config = require('./../config')
-const { projectDir, starterDir, appDir } = require('../path-utils')
+const { projectDir, starterDir, appDir } = require('../utils/paths')
 
 const migrationSteps = require('./migration-steps')
 const { preflightChecks, deleteIfUnchanged, removeOldPatternIncludesFromSassFile } = require('./migration-steps')
