+++ lib/plugins/plugins.js
@@ -40,7 +40,7 @@ const path = require('path')
 
 // Local dependencies
 const appConfig = require('../config')
-const { projectDir } = require('../path-utils')
+const { projectDir } = require('../utils/paths')
 const pkgPath = path.join(projectDir, 'package.json')
 const chokidar = require('chokidar')
 const fse = require('fs-extra')
