+++ lib/plugins/plugins.test.js
@@ -6,7 +6,7 @@ const config = require('../config')
 
 // Local dependencies
 const plugins = require('./plugins')
-const fakeFileSystem = require('../../__tests__/util/mock-file-system')
+const fakeFileSystem = require('../../__tests__/utils/mock-file-system')
 
 // Local variables
 const rootPath = path.join(__dirname, '..', '..')
