+++ lib/routes/prototype-admin-routes.js
@@ -7,7 +7,7 @@ const contextPath = '/manage-prototype'
 const router = require('../../index').requests.setupRouter(contextPath)
 const redirectingRouter = require('../../index').requests.setupRouter('/manage-prototype')
 
-const { projectDir, packageDir } = require('../path-utils')
+const { projectDir, packageDir } = require('../utils/paths')
 const path = require('path')
 const plugins = require('../plugins/plugins')
 const config = require('../config').getConfig()
