+++ lib/usage-data.js
@@ -10,7 +10,7 @@ const { v4: uuidv4 } = require('uuid')
 
 // Local dependencies
 const packageJson = require('../package.json')
-const { projectDir } = require('./path-utils')
+const { projectDir } = require('./utils/paths')
 const usageDataFilePath = path.join(projectDir, 'usage-data-config.json')
 
 exports.getUsageDataConfig = function () {
