+++ lib/utils/index.js
@@ -10,11 +10,11 @@ const portScanner = require('portscanner')
 const { marked } = require('marked')
 
 // Local dependencies
-const config = require('./config').getConfig()
-const plugins = require('./plugins/plugins')
-const filters = require('./filters/api')
-const routes = require('./routes/api')
-const { appDir, tempDir, projectDir } = require('./path-utils')
+const config = require('../config').getConfig()
+const plugins = require('../plugins/plugins')
+const filters = require('../filters/api')
+const routes = require('../routes/api')
+const { appDir, tempDir, projectDir } = require('./paths')
 const { existsSync } = require('fs')
 
 // Tweak the Markdown renderer
