+++ lib/utils/session.test.js
@@ -2,7 +2,7 @@
 
 const nunjucks = require('nunjucks')
 
-const sessionUtils = require('./session-utils.js')
+const sessionUtils = require('./session.js')
 
 describe('checked', () => {
   let ctx, checked
