+++ lib/utils/utils.test.js
@@ -11,7 +11,7 @@ describe('sessionFileStoreQuietLogFn', () => {
 
     const session = require('express-session')
     const SessionFileStore = require('session-file-store')(session)
-    const { sessionFileStoreQuietLogFn } = require('./utils')
+    const { sessionFileStoreQuietLogFn } = require('./index')
 
     jest.useFakeTimers({ doNotFake: ['performance'] })
